heyo

+++ srcs/ft_nbr.c
@@ -6,7 +6,7 @@
 /*   By: rde-kwaa <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/07/09 12:18:46 by rde-kwaa          #+#    #+#             */
-/*   Updated: 2018/07/21 12:49:10 by rde-kwaa         ###   ########.fr       */
+/*   Updated: 2018/07/21 13:17:14 by rde-kwaa         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -20,7 +20,7 @@ int		ft_nbr(va_list arg, char f)
 	int		len;
 
 	n = va_arg(arg, int);
-	len = ft_flags(f, 'd', (long)n);
+	len = ft_flags(f, 'd', (long long)n);
 	s = ft_itoa(n);
 	ft_putstr(s);
 	len += ft_strlen(s);
